+++ src/main/kotlin/io/kjson/uri/VariableReference.kt
@@ -1,0 +1,7 @@
+package io.kjson.uri
+
+class VariableReference(
+    val variable: Variable,
+    val characterLimit: Int?,
+    val explode: Boolean,
+)